requirements.txt
asgiref==3.6.0
black==22.12.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Django==4.1.5
django-cors-headers==3.13.0
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
idna==3.4
mypy-extensions==0.4.3
pathspec==0.10.3
phonenumbers==8.13.4
platformdirs==2.6.2
PyJWT==2.6.0
pyotp==2.8.0
python-decouple==3.7
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
twilio==7.16.1
urllib3==1.26.14
